+++ tests/test_transformation_lagrange.py
@@ -14,7 +14,7 @@ import pytest
 
 from numpy.linalg import LinAlgError
 
-from conftest import build_rnd_coeffs, create_non_downward_closed_multi_index
+from conftest import build_rnd_coeffs
 
 from minterpy import (
     MultiIndexSet,
@@ -101,7 +101,7 @@ class TestNonDownwardClosed:
       for the transformation. Furthermore, certain transformations strictly
       require that the polynomials having a downward-close multi-index set.
     """
-    def test_to_newton(self, SpatialDimension, PolyDegree, LpDegree):
+    def test_to_newton(self, multi_index_non_downward_closed):
         """Test the transformation to the Newton basis.
 
         Notes
@@ -109,12 +109,8 @@ class TestNonDownwardClosed:
         - If the multi-index set is not downward-closed, the naive
           transformation operator is automatically selected.
         """
-        # Create a non-downward-closed set
-        mi = create_non_downward_closed_multi_index(
-            SpatialDimension,
-            PolyDegree,
-            LpDegree
-        )
+        # Get the non-downward-closed set
+        mi = multi_index_non_downward_closed
         assert not mi.is_downward_closed
 
         # Create a Lagrange polynomial
@@ -131,7 +127,7 @@ class TestNonDownwardClosed:
         # Assertion
         assert np.allclose(lag_coeffs_ref, lag_coeffs)
 
-    def test_to_canonical(self, SpatialDimension, PolyDegree, LpDegree):
+    def test_to_canonical(self, multi_index_non_downward_closed):
         """Test the transformation to the canonical basis.
 
         Notes
@@ -139,12 +135,8 @@ class TestNonDownwardClosed:
         - The transformation to the canonical basis strictly requires that
           the multi-index set is downward-closed.
         """
-        # Create a non-downward-closed set
-        mi = create_non_downward_closed_multi_index(
-            SpatialDimension,
-            PolyDegree,
-            LpDegree
-        )
+        # Get the non-downward-closed set
+        mi = multi_index_non_downward_closed
         assert not mi.is_downward_closed
 
         # Create a Lagrange polynomial
@@ -155,7 +147,7 @@ class TestNonDownwardClosed:
         with pytest.raises(ValueError):
             LagrangeToCanonical(lag_poly)()
 
-    def test_to_chebyshev(self, SpatialDimension, PolyDegree, LpDegree):
+    def test_to_chebyshev(self, multi_index_non_downward_closed):
         """Test the transformation to the Chebyshev basis.
 
         Notes
@@ -167,11 +159,7 @@ class TestNonDownwardClosed:
           caught below.
         """
         # Create a non-downward-closed set
-        mi = create_non_downward_closed_multi_index(
-            SpatialDimension,
-            PolyDegree,
-            LpDegree
-        )
+        mi = multi_index_non_downward_closed
         assert not mi.is_downward_closed
 
         # Create a Lagrange polynomial